@@ -1,6 +1,6 @@
 module open-cluster-management.io/cert-policy-controller
 
-go 1.23.0
+go 1.24.0
 
 require (
 	github.com/go-logr/zapr v1.3.0
